Provide `AESECBCmd` command alternatives using byte arrays.
In cryptography operations are usually done on byte arrays. Java `String`s
do use bytes underneath, but they also add character encoding information.
So when converting between a byte array and a `String` a certain risk
exists that the bytes are altered during the conversions.

For this reason, and in order to prevent the unnecessary conversion to
and from strings for API consumers that operate directly on byte arrays, an
alternative version for all commands in `AESECDCmd` is added that passes
through the byte arrays.

All the methods using `String` are still left in, of course.